requirements.txt
Flask
Werkzeug
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
flask-cors
google-genai
firebase-admin
gunicorn